# PERF: Improve performance for MultiIndex.isin