- [ ] closes #xxxx (Replace xxxx with the Github issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

```
       before           after         ratio
     [85246fe4]       [d6015d67]
     <sym_diff^2>       <isin_performance>
-     3.48±0.01ms       3.13±0.1ms     0.90  multiindex_object.Isin.time_isin('string')
-     3.41±0.02ms      1.46±0.02ms     0.43  multiindex_object.Isin.time_isin('int')
-     10.2±0.04ms      1.54±0.02ms     0.15  multiindex_object.Isin.time_isin('datetime')
```